refactor(store): modularize tensorcast.api.store into subpackages for improved clarity and maintainability
